+++ plom/client/tools/tick.py
@@ -9,6 +9,7 @@ from PyQt6.QtWidgets import QGraphicsPathItem, QGraphicsItem
 
 from plom.client.tools import OutOfBoundsPen, OutOfBoundsFill
 from plom.client.tools import CommandTool, UndoStackMoveMixin
+from plom.client.tools import DefaultTickRadius
 
 
 class CommandTick(CommandTool):
@@ -28,18 +29,17 @@ class CommandTick(CommandTool):
 
 
 class TickItem(UndoStackMoveMixin, QGraphicsPathItem):
-    tick_radius = 20
-
     def __init__(self, pt, style):
         super().__init__()
+        tick_radius = style["scale"] * DefaultTickRadius
         self.saveable = True
         self.pt = pt
-        self.path = QPainterPath()
+        path = QPainterPath()
         # Draw the checkmark with barycentre under mouseclick.
-        self.path.moveTo(pt.x() - self.tick_radius / 2, pt.y() - self.tick_radius / 2)
-        self.path.lineTo(pt.x(), pt.y())
-        self.path.lineTo(pt.x() + self.tick_radius, pt.y() - self.tick_radius)
-        self.setPath(self.path)
+        path.moveTo(pt.x() - tick_radius / 2, pt.y() - tick_radius / 2)
+        path.lineTo(pt.x(), pt.y())
+        path.lineTo(pt.x() + tick_radius, pt.y() - tick_radius)
+        self.setPath(path)
         self.restyle(style)
 
         self.setFlag(QGraphicsItem.GraphicsItemFlag.ItemIsMovable)